move signedDistanceToMesh
