+++ source/MRMesh/MRMeshDistance.cpp
@@ -2,10 +2,49 @@
 #include "MRAABBTree.h"
 #include "MRMesh.h"
 #include "MRTriDist.h"
+#include "MRLine3.h"
+#include "MRMeshIntersect.h"
 #include "MRTimer.h"
 
 namespace MR
 {
+
+std::optional<float> signedDistanceToMesh( const MeshPart& mp, const Vector3f& p, const DistanceToMeshOptions& op )
+{
+    assert( op.signMode != SignDetectionMode::OpenVDB );
+    const auto proj = findProjection( p, mp, op.maxDistSq, nullptr, op.minDistSq );
+    if ( op.signMode != SignDetectionMode::HoleWindingRule // for HoleWindingRule the sign can change even for too small or too large distances
+        && ( proj.distSq < op.minDistSq || proj.distSq >= op.maxDistSq ) ) // note that proj.distSq == op.minDistSq (e.g. == 0) is a valid situation
+        return {}; // distance is too small or too large, discard them
+
+    float dist = std::sqrt( proj.distSq );
+    switch ( op.signMode )
+    {
+    case SignDetectionMode::ProjectionNormal:
+        if ( !mp.mesh.isOutsideByProjNorm( p, proj, mp.region ) )
+            dist = -dist;
+        break;
+
+    case SignDetectionMode::WindingRule:
+    {
+        const Line3d ray( Vector3d( p ), Vector3d::plusX() );
+        int count = 0;
+        rayMeshIntersectAll( mp, ray, [&count] ( auto&& ) { ++count; return true; } );
+        if ( count % 2 == 1 ) // inside
+            dist = -dist;
+        break;
+    }
+
+    case SignDetectionMode::HoleWindingRule:
+        assert( !mp.region );
+        if ( !mp.mesh.isOutside( p, op.windingNumberThreshold, op.windingNumberBeta ) )
+            dist = -dist;
+        break;
+
+    default: ; //nothing
+    }
+    return dist;
+}
 
 void processCloseTriangles( const MeshPart& mp, const Triangle3f & t, float rangeSq, const TriangleCallback & call )
 {
